Expand login tests to cover all key projects and fix config
