+++ .env.playwright
@@ -1,1 +1,23 @@
-BASE_URL=https://localhost:3000
+PROJECTS='{
+    "atlas":{
+        "url": "https://atlas.kontur.io/",
+        "app": "9043acf9-2cf3-48ac-9656-a5d7c4b7593d",
+        "title":
+        "Kontur Atlas"
+    },
+    "disaster_ninja":{
+        "url":"https://disaster.ninja/",
+        "app":"58851b50-9574-4aec-a3a6-425fa18dcb54",
+        "title":"Disaster Ninja"
+    },
+    "smart_city":{
+        "url":"https://maps.kontur.io/",
+        "app":"634f23f5-f898-4098-a8bd-09eb7c1e1ae5",
+        "title":"Smart City"
+    },
+    "oam":{
+        "url":"https://new.openaerialmap.org/",
+        "app":"1dc6fe68-8802-4672-868d-7f17943bf1c8",
+        "title":"OpenAerialMap"
+    }
+}'
